requirements.txt
psycopg2
dj-database-url
gunicorn
djangorestframework
djangorestframework-simplejwt

certifi==2022.9.24
crispy-bootstrap5==0.7
dj-database-url==1.3.0
Django==4.1.7
django-crispy-forms==2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
numpy==1.24.0
pandas==1.5.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
typing_extensions==4.5.0
